requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.2
